updated aws sdk to v2, code organization
